requirements.txt
tweepy==4.14.0
google-generativeai==0.3.2
schedule==1.2.0
